requirements.txt
pycryptodome>=3.20.0
requests>=2.32.3
eth-account>=0.10.0
python-dotenv>=1.0.1
